Experiment with one method of form validation
This is one way to automatically handle form parameter validation for
web routes. It works the same as the API routes where the Malli library
is used as a middleware to coerce a request.

The difference for the web routes is that a custom exception handler
middleware is used to catch errors thrown by Malli. At this point the
route metadata is checked for a :view-fn key which specifies the
function that should be called to provide the response HTML for the page
being viewed.

The idea is to either (a) provide this view function with the exception
details so they can be added to the form by matching on field names, or
(b) add the exception message to the flash.

This is only one attempt at. Another approach, which will be explored
later, would be to not perform this request coercion as a middleware but
to call the Malli validate function manually in the route
handler/controller. If this approach is chosen it would be nice to
ensure the API routes also work the same way.